runpod-worker-wan/requirements.txt
# === Wan2.1 I2V Worker — зависимости ===

# RunPod
runpod

# Утилиты
python-dotenv
boto3
Pillow
requests

# PyTorch и diffusers ставятся в Dockerfile отдельно (с CUDA)
# accelerate нужен для enable_model_cpu_offload
accelerate
safetensors
transformers>=4.40.0
imageio
imageio-ffmpeg
